+++ test/readonly.ts
@@ -18,6 +18,14 @@ describe('readonly', () => {
     T.decode({ a: 1 }).map(x => assert.ok(Object.isFrozen(x)))
   })
 
+  it('should not freeze in production', () => {
+    const env = process.env.NODE_ENV
+    process.env.NODE_ENV = 'production'
+    const T = t.readonly(t.interface({ a: t.number }))
+    T.decode({ a: 1 }).map(x => assert.ok(!Object.isFrozen(x)))
+    process.env.NODE_ENV = env
+  })
+
   it('should serialize a deserialized', () => {
     const T = t.readonly(t.interface({ a: DateFromNumber }))
     assert.deepEqual(T.encode({ a: new Date(0) }), { a: 0 })
@@ -37,5 +45,12 @@ describe('readonly', () => {
     assert.strictEqual(T2.is({ a: new Date(0) }), true)
     assert.strictEqual(T2.is({ a: 0 }), false)
     assert.strictEqual(T2.is(undefined), false)
+  })
+
+  it('should assign a default name', () => {
+    const T1 = t.readonly(t.type({ a: t.number }))
+    assert.strictEqual(T1.name, 'Readonly<{ a: number }>')
+    const T2 = t.readonly(t.type({ a: t.number }), 'T2')
+    assert.strictEqual(T2.name, 'T2')
   })
 })