+++ test/taggedUnion.ts
@@ -30,7 +30,19 @@ const TUC = t.exact(
   'TUC'
 )
 
-const T = t.taggedUnion('type', [TUA, TUB, TUC])
+const TUD = t.intersection(
+  [
+    t.type({
+      bar: t.number
+    }),
+    t.type({
+      type: t.literal('d')
+    })
+  ],
+  'TUD'
+)
+
+const T = t.taggedUnion('type', [TUA, TUB, TUC, TUD])
 
 describe('taggedUnion', () => {
   it('should succeed validating a valid value', () => {
@@ -45,16 +57,18 @@ describe('taggedUnion', () => {
   })
 
   it('should fail validating an invalid value', () => {
-    assertFailure(T.decode(true), ['Invalid value true supplied to : (TUA | TUB | TUC)'])
-    assertFailure(T.decode({ type: 'D' }), ['Invalid value "D" supplied to : (TUA | TUB | TUC)/type: "a" | "b" | "c"'])
+    assertFailure(T.decode(true), ['Invalid value true supplied to : (TUA | TUB | TUC | TUD)'])
+    assertFailure(T.decode({ type: 'D' }), [
+      'Invalid value "D" supplied to : (TUA | TUB | TUC | TUD)/type: "a" | "b" | "c" | "d"'
+    ])
     assertFailure(T.decode({ type: 'a' }), [
-      'Invalid value undefined supplied to : (TUA | TUB | TUC)/0: TUA/foo: string'
+      'Invalid value undefined supplied to : (TUA | TUB | TUC | TUD)/0: TUA/foo: string'
     ])
     assertFailure(T.decode({ type: 'b' }), [
-      'Invalid value undefined supplied to : (TUA | TUB | TUC)/1: TUB/bar: number'
+      'Invalid value undefined supplied to : (TUA | TUB | TUC | TUD)/1: TUB/bar: number'
     ])
     assertFailure(T.decode({ type: 'c' }), [
-      'Invalid value undefined supplied to : (TUA | TUB | TUC)/2: TUC/baz: DateFromNumber'
+      'Invalid value undefined supplied to : (TUA | TUB | TUC | TUD)/2: TUC/baz: DateFromNumber'
     ])
   })
 
@@ -182,5 +196,37 @@ describe('taggedUnion', () => {
     assertSuccess(U.decode({ type: '1', bar: 2 }))
     assertFailure(U.decode({ type: '1', foo: 'foo' }), ['Invalid value undefined supplied to : U/1: B/bar: number'])
     assertFailure(U.decode({ type: 3 }), ['Invalid value 3 supplied to : U/type: 1 | "1"'])
+  })
+})
+
+describe('isTagged', () => {
+  const is = t.isTagged('type')
+  it('should return true if the type is tagged', () => {
+    const T1 = t.type({ type: t.literal('a') })
+    const T2 = t.strict({ type: t.literal('b') })
+    assert.ok(is(T1))
+    assert.ok(is(T2))
+    assert.ok(is(t.union([T1, T2])))
+    assert.ok(is(t.intersection([T1, t.type({ a: t.string })])))
+    assert.ok(is(t.refinement(T1, () => true)))
+    assert.ok(is(t.exact(T1)))
+  })
+
+  it('should return false if the type is not tagged', () => {
+    assert.ok(!is(t.string))
+  })
+})
+
+describe('getTagValue', () => {
+  const get = t.getTagValue('type')
+  it('should return the tag value', () => {
+    const T1 = t.type({ type: t.literal('a') })
+    const T2 = t.strict({ type: t.literal('b') })
+    assert.strictEqual(get(T1), 'a')
+    assert.strictEqual(get(T2), 'b')
+    assert.strictEqual(get(t.union([T1, T1])), 'a')
+    assert.strictEqual(get(t.intersection([T1, t.type({ a: t.string })])), 'a')
+    assert.strictEqual(get(t.refinement(T1, () => true)), 'a')
+    assert.strictEqual(get(t.exact(T1)), 'a')
   })
 })